chgres_grib2:  This commit references #63782.
Bug fix for calculation of grid corners for lambert conformal grid where
grid size was being broadcast as an integer instead of a real8,
resulting in improper corner location computation.

Added a short to soil moisture rescaling to prevent the routine from
running if soil moisture is missing in the input data.

Small syntax fixes for input_data.F90