chgres_grib2:  This commit references #63782 Makefile library order fix
In loading the wgrib2_api and wgrib2_lib libraries before the nemsio
libraries, there was a conflict between the splat routine in nemsio and
the included ipolates splat routine from wgrib2. This was resulting in
incorrect latitude arrays for gaussian files. The order of the libraries
loaded in the makefile was changed so that the nemsio version of splat
(i.e., the one the code expects) is used. If wgrib2 is built without
ipolates, which is not needed for this code, then this conflift doesn't
exist.