feature/hera_port  This commit references #1.
Point to George V's copy of the nesmio library instead of
the 'official' version.  The nemsio utilities did not
work with the 'official' version - they return bad istat (-41)
when reading.  Either the official version was compiled wrong
or as a newer version (v2.5) it may be incompatible with the
programs.