support/HAFS: Enable make_hgrid shift_fac option for uniform grid.
Regarding the center longitude of the uniform global grid, the default center
longitude for tile 1 was hardwired at 10.0W, even though in the source code
instructions there is a commandline option (--shift_fac), which can shift the
center longitude. For example:
In make_hgrid.c:
...
  "   --shift_fac                shift west by 180/shift_fac. Default value is 18.  ",
...

In the source code, this shift_fac option, was hard-wired as 18.0
(corresponding to setting the tile 1 center longtintude to 10.0W).

In this commit, the capability of using this shift_fac option is enabled.
Thus, if one wants to specify the center lon for tile 1 to be 60.0W, he can
specify --shift_fac=3.0 (180./3.0=60, meaning shift 60 degrees westward).

In any case, the default value of shift_fac is still 18.0, thus the default
tile 1 center longitude is still 10.0W. This commit will not change the default
behavior, rather than provide the capability/flexibility to cumstomize the tile
1 center longitude if that is desired.

A note from George Gyno: I think GFDL hardwired it to ensure tall mountains
were located away from the tile edges.