CHGRES_CUBE: range check SST when using GFS GRIB2 input data
Add range check to remove bad SST values (< 271.2K or > 310K) when using v14, v15.1 and v15.2 GFS GRIB2 data.  Bad values are the result of bi-linear interpolation from the gaussian master files to the lat/lon "pgb" files.   As a result, coastal points (even those identified as water by the mask) can be a mixture of land and non-land.   

This also fixes a model failure discovered when testing v1 of the public release.

See issue #59 for more details.