release/ops-gefsv12.0.1  This commit references #62.
Update per GEFS group to support v12.

An option, DOSFC, is added that allows the surface/nst regridding
to be skipped. The default is "YES" to maintain so the same
functionality is maintained. This is added to improve the efficiency
of some GEFS scripts.

Additionally, a new file, chgres_atm.log is created in OUTDIR to
signify the atmosphere portion is complete. This is also to aid in
GEFS efficiency.