Fix bug in chgres_cube that prevents Thompson aerosols from being created
The return statement in read_setup_namelist is a few lines too early, and 
before the if statement that sets use_thomp_mp_climo to true if 
thomp_mp_climo_file is set in the namelist. This prevents the aerosols from 
ever being created or written to file, regardless of input data type. 

Addresses issue #178