Update global_cycle program to apply soil temperature increments from GSI. (#420)
New feature to read soil temperature increments from the GSI, then apply 
the increments to the model states.  Feature is only active when the
DO_LNDINC logical is set.

Includes a new global_cycle consistency test for this option.

Includes a new unit test to check the algorithm that adjusts the 
soil temperature and the ratio of frozen/liquid soil moisture.

Fixes #300.