global_cycle - New option to apply snow depth increments from JEDI (#568)
Addition of routines to add snow depth increments from the JEDI output 
increment file to the surface restart files. Includes screening to apply 
updates over land only and to adjust the SWE to match the snow depth analysis. 

This is an option. The default is to not read in or apply the increments.

Fixes #561