Update build modules to be lua compliant (#614)
Update build modules for Hera, Jet, Orion and WCOSS-Dell to be "lua" compliant.
This is required on our new WCOSS2 machine. So update the rest of the repository
accordingly. Allow default library versions to be overridden by environment variables.

Update to ESMF v8.1.1 on Hera, because the version that was being used (ESMF v8.1.0 beta
snapshot 27) was removed without warning.

Minor consistency test script updates to prevent a namelist read error when
using the GNU compiler.

Fixes #613