Add TEST_FILE_DIR option to CMake build, where test data files can be found instead of using FTP.  (#732)
Add handling for the TEST_FILE_DIR option in CMake. This allows the programmer to 
specify a local directory to check first, before downloading test data files from the FTP site.

Use this capability, and a cache, to cache the data for all CI runs as well. This prevents
the CI system from downloading these test files hundreds of times per day.

Part of #719.
Fixes #728