Remove goto statements from chgres_cube (#775)
The WAM option was using the NRLMSISE-00 library, which contains numerous 'goto'
statements. That library was replaced with the latest NRLMSIS 2.1 library, which 
does not use 'goto' statements and is also a welcome improvement in initializing 
the upper atmosphere. 

The new library included a unit test, which was added to run under Github
actions.

Fixes #759.