sfc_climo_gen - Output fractions of each vegetation/soil type category (#748)
The sfc_climo_gen program was updated to optionally output the fraction
of each vegetation and soil type category at a model point.  The dominant
category is also output. The default is to output just the dominant category.

Unit and consistency tests were added to test this new capability.

The 'readthedocs' was also updated.

Fixes #709.