Add backup calculation for orography and mask generation (#713)
For model grids approaching 1 km, the orography and mask generation can fail when there
are no source data points contained within a model grid box.  In this situation, a backup 
calculation is used that considers source data points immediately surrounding the model grid box.

This is meant to be a temporary fix until high-resolution source data is added to
the orography code.

Fixes #714.