Build flexibility. (#836)
Flexibility to call the "build_all.sh" script from anywhere, thereby relaxing 
the need to "cd" into the cloned ufs_utils directory. 

Option to specify a build directory (the default is retained as ./build).

Option to reuse a previously created build directory (the default is to create a 
new "./build" directory with every execution of "build_all.sh" as was before).

Option to provide an installation prefix (the default is retained as the current directory).