global_cycle - Removed checks preventing land increments for fractional grids. (#903)
Removed two checks that caused global_cycle to report an error and exit in the 
case that both land increments were being added and fractional grids are used. There 
are no issues adding soil moisture and soil temperature increments in fractional grid mode, 
so the code can be allowed to proceed. 

Fixes #895.