chgres_cube: Add a check for incompatible surface static data (#911)
Add logic to ensure that the orogoraphy data is consistent with the static climatological
surface data. The check looks for any points where land fraction > 0 but the static data
point is missing (-999 flag). This will happen with older version of the orography data.
The program now fails if incompatible static data is detected.

Fixes #906.