Update 'readthedocs' for new AWS fixed files host (#927)
Update the 'readthedocs' chgres_cube.rst file as follows: Link to the 
global workflow group's AWS bucket for the global model 'fixed' files 
('grid', mosaic, 'oro' and surface). Update the text for the new 'oro' and surface 
file naming convention, which uses the ocean resolution in the file name.

Fixes #926.