chgres_cube - modify interpolation of canopy moisture content (#1053)
Update interpolation from 'all land' to 'all land' to 'land-no ice' to 'land-no-ice'.
This prevents the large flag values at land ice points in the GFS v17 history files
from being interpolated to the target grid. As a consequence of this update,
the interpolation method was changed from conservative to neighbor.

Fixes #1052