regrid_sfc - Add a time dimension to output files (#1063)
Modify the surface regridding (regrid_sfc) code to add a time dimension to the output 
files based on a namelist option. The previous version added the time dimension only 
to files with more than one timestep. However, the UFS model expects the time dimension for all files.

Also add a namelist template that is used by the global workflow.