ocnice_prep - Allow warmstart IC creation from low->high resolution (#1061)
The ocnice_prep utility was modified to allow generation of MOM6 and CICE6 warmstart files 
from lower resolution restart files. And to allow additional source and destination grids.
The cpld_gridgen was modified to create the required weights if the destination grid is 1/4 degree

Minor updates to the documentation of both utilities were made to reflect the changes.

Fixes #1059